requirements.txt
Flask>=2.0
google-genai>=1.69.0
google-generativeai>=0.1
